Handle rescheduling of flushed diagnostics
When diagnostics are flushed after leaving insert mode, it's possible
the user entered insert mode again before the diagnostics are displayed.

This commits ensures that if we're about to display flushed diagnostics,
we don't do so if the user entered insert mode again. This way if you
make some changes, exit insert mode and enter it again, the diagnostics
are not displayed until you exit insert mode again _and_ wait for the
timer to expire.