Improve buffer version checking
Instead of comparing the current buffer version to the LSP document
version, we now compare the current buffer version to the version before
diagnostics were scheduled for presentation. This ensures that if a
buffer is formatted (and thus changed) _after_ sending a didChange but
_before_ diagnostics came in, we still show the diagnostics.

Getting all of this right is pretty tricky, and I suspect this new setup
won't be perfect either. Unfortunately, I think this is the best we can
do to eliminate flickering of underlines/signs/etc.