Remove buffer version checking
As expected, this turns out to be impossible to get right. Even with the
previous implementation, occasionally a buffer would change between
diagnostics being deferred and displayed. This was especially apparent
with rust-analyzer, as most of its diagnostics are not produced until
you write the file, meaning outdated diagnostics can stick around for a
while.

So we're back to just not caring about buffer/diagnostic document
versions, as this seems to be the least annoying approach of all.