Fix memory leak when publishing diagnostics
For reasons unknown, the closure used for vim.defer_fn() is retained for
a potentially long amount of time, even after calling collectgarbage().
Since this closure captures a large amount of data, that data too stays
around for long. This in turn can result in a memory leak when
processing many (e.g. 300+ diagnostics).

To resolve this, the data is stored/kept in the cache, and the
schedule() closure reads from that cache and clears it. This way we can
achieve the same results but without having to capture any large data.